Disallow `force: true` when creating table
The `force: true` option to `ActiveRecord::Migration#create_table`
causes the table to be dropped if it already exists. This is unsafe by
definition, and should instead require an explicit `drop_table` instead.
To ensure we don't break backwards compatibility, I've configured the
default for this new feature to continue to allow the old behavior --
making it opt-in for now, but with the expectation that we'll flip the
default in the next major version.